Added support for all unsigned primitive types.
- tried to implement an umbrella trait for them, but for some reason,
  that didn't work.
- it might be worth investigating why tho
